pyproject.toml
[tool.poetry]
name = "gpt-wrap"
version = "0.1.0"
description = "Teregram bot wrapped around GPT API"
authors = ["mrYush"]
license = "MIT License"
readme = "README.md"
packages = [{include = "gpt_wrap"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.2"
pyyaml = "^6.0"
python-telegram-bot = "^20.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
